+++ packages/node/src/sync/stream/NodeRemote.ts
@@ -11,9 +11,15 @@ import {
   RemoteConnector
 } from '@powersync/common';
 import { BSON } from 'bson';
-import Agent from 'proxy-agent';
-import { EnvHttpProxyAgent, Dispatcher } from 'undici';
-import { WebSocket } from 'ws';
+import {
+  Dispatcher,
+  EnvHttpProxyAgent,
+  ErrorEvent,
+  getGlobalDispatcher,
+  ProxyAgent,
+  WebSocket as UndiciWebSocket
+} from 'undici';
+import { ErrorRecordingDispatcher } from './ErrorRecordingDispatcher.js';
 
 export const STREAMING_POST_TIMEOUT_MS = 30_000;
 
@@ -23,36 +29,77 @@ class NodeFetchProvider extends FetchImplementationProvider {
   }
 }
 
-export type NodeRemoteOptions = AbstractRemoteOptions & {
+export type NodeCustomConnectionOptions = {
+  /**
+   * Optional custom dispatcher for HTTP or WEB_SOCKET connections.
+   *
+   * This can be used to customize proxy usage (using undici ProxyAgent),
+   * or other connection options.
+   */
   dispatcher?: Dispatcher;
 };
 
+export type NodeRemoteOptions = AbstractRemoteOptions & NodeCustomConnectionOptions;
+
 export class NodeRemote extends AbstractRemote {
+  private wsDispatcher: Dispatcher | undefined;
+
   constructor(
     protected connector: RemoteConnector,
     protected logger: ILogger = DEFAULT_REMOTE_LOGGER,
     options?: Partial<NodeRemoteOptions>
   ) {
-    // EnvHttpProxyAgent automatically uses relevant env vars for HTTP
-    const dispatcher = options?.dispatcher ?? new EnvHttpProxyAgent();
+    const fetchDispatcher = options?.dispatcher ?? defaultFetchDispatcher();
 
     super(connector, logger, {
       fetchImplementation: options?.fetchImplementation ?? new NodeFetchProvider(),
       fetchOptions: {
-        dispatcher
+        dispatcher: fetchDispatcher
       },
       ...(options ?? {})
     });
+
+    this.wsDispatcher = options?.dispatcher;
   }
 
   protected createSocket(url: string): globalThis.WebSocket {
-    return new WebSocket(url, {
-      // Automatically uses relevant env vars for web sockets
-      agent: new Agent.ProxyAgent(),
+    // Create dedicated dispatcher for this WebSocket
+    const baseDispatcher = this.getWebsocketDispatcher(url);
+    const errorRecordingDispatcher = new ErrorRecordingDispatcher(baseDispatcher);
+
+    // Create WebSocket with dedicated dispatcher
+    const ws = new UndiciWebSocket(url, {
+      dispatcher: errorRecordingDispatcher,
       headers: {
         'User-Agent': this.getUserAgent()
       }
-    }) as any as globalThis.WebSocket; // This is compatible in Node environments
+    });
+
+    errorRecordingDispatcher.onError = (error: Error) => {
+      // When we receive an error from the Dispatcher, emit the event on the websocket.
+      // This will take precedence over the WebSocket's own error event, giving more details on what went wrong.
+      const event = new ErrorEvent('error', {
+        error,
+        message: error.message
+      });
+      ws.dispatchEvent(event);
+    };
+
+    return ws as globalThis.WebSocket;
+  }
+
+  protected getWebsocketDispatcher(url: string) {
+    if (this.wsDispatcher != null) {
+      return this.wsDispatcher;
+    }
+
+    const protocol = new URL(url).protocol.replace(':', '');
+    const proxy = getProxyForProtocol(protocol);
+    if (proxy != null) {
+      return new ProxyAgent(proxy);
+    } else {
+      return getGlobalDispatcher();
+    }
   }
 
   getUserAgent(): string {
@@ -67,4 +114,22 @@ export class NodeRemote extends AbstractRemote {
   async getBSON(): Promise<BSONImplementation> {
     return BSON;
   }
+}
+
+function defaultFetchDispatcher(): Dispatcher {
+  // EnvHttpProxyAgent automatically uses HTTP_PROXY, HTTPS_PROXY and NO_PROXY env vars by default.
+  // We add ALL_PROXY support.
+  return new EnvHttpProxyAgent({
+    httpProxy: getProxyForProtocol('http'),
+    httpsProxy: getProxyForProtocol('https')
+  });
+}
+
+function getProxyForProtocol(protocol: string): string | undefined {
+  return (
+    process.env[`${protocol.toLowerCase()}_proxy`] ??
+    process.env[`${protocol.toUpperCase()}_PROXY`] ??
+    process.env[`all_proxy`] ??
+    process.env[`ALL_PROXY`]
+  );
 }