+++ tools/diagnostics-app/src/library/powersync/ConnectionManager.ts
@@ -1,20 +1,19 @@
 import {
   BaseListener,
-  BaseObserver,
+  createBaseLogger,
+  LogLevel,
   PowerSyncDatabase,
+  TemporaryStorageOption,
+  WASQLiteOpenFactory,
+  WASQLiteVFS,
   WebRemote,
   WebStreamingSyncImplementation,
-  WebStreamingSyncImplementationOptions,
-  WASQLiteOpenFactory,
-  TemporaryStorageOption,
-  WASQLiteVFS,
-  createBaseLogger,
-  LogLevel
+  WebStreamingSyncImplementationOptions
 } from '@powersync/web';
+import { safeParse } from '../safeParse/safeParse';
 import { DynamicSchemaManager } from './DynamicSchemaManager';
 import { RecordingStorageAdapter } from './RecordingStorageAdapter';
 import { TokenConnector } from './TokenConnector';
-import { safeParse } from '../safeParse/safeParse';
 
 const baseLogger = createBaseLogger();
 baseLogger.useDefaults();
@@ -45,47 +44,13 @@ export const db = new PowerSyncDatabase({
 
 export const connector = new TokenConnector();
 
-const remote = new WebRemote(connector);
 const adapter = new RecordingStorageAdapter(db.database, schemaManager);
 
-const syncOptions: WebStreamingSyncImplementationOptions = {
-  adapter,
-  remote,
-  uploadCrud: async () => {
-    // No-op
-  },
-  identifier: 'diagnostics'
-};
-export const sync = new WebStreamingSyncImplementation(syncOptions);
+export let sync: WebStreamingSyncImplementation | undefined;
 
 export interface SyncErrorListener extends BaseListener {
   lastErrorUpdated?: ((error: Error) => void) | undefined;
 }
-
-class SyncErrorTracker extends BaseObserver<SyncErrorListener> {
-  public lastSyncError: Error | null = null;
-
-  constructor() {
-    super();
-    // Big hack: Use the logger to get access to connection errors
-    const defaultHandler = baseLogger.createDefaultHandler();
-    baseLogger.setHandler((messages, context) => {
-      defaultHandler(messages, context);
-      if (context.name == 'PowerSyncStream' && context.level.name == 'ERROR') {
-        if (messages[0] instanceof Error) {
-          this.lastSyncError = messages[0];
-        } else {
-          this.lastSyncError = new Error('' + messages[0]);
-        }
-        this.iterateListeners((listener) => {
-          listener.lastErrorUpdated?.(this.lastSyncError!);
-        });
-      }
-    });
-  }
-}
-
-export const syncErrorTracker = new SyncErrorTracker();
 
 if (connector.hasCredentials()) {
   connect();
@@ -93,33 +58,44 @@ if (connector.hasCredentials()) {
 
 export async function connect() {
   const params = getParams();
+  await sync?.disconnect();
+  const remote = new WebRemote(connector);
+  const syncOptions: WebStreamingSyncImplementationOptions = {
+    adapter,
+    remote,
+    uploadCrud: async () => {
+      // No-op
+    },
+    identifier: 'diagnostics'
+  };
+  sync = new WebStreamingSyncImplementation(syncOptions);
   await sync.connect({ params });
   if (!sync.syncStatus.connected) {
+    const error = sync.syncStatus.dataFlowStatus.downloadError ?? new Error('Failed to connect');
     // Disconnect but don't wait for it
-    sync.disconnect();
-    throw syncErrorTracker.lastSyncError ?? new Error('Failed to connect');
-  } else {
-    syncErrorTracker.lastSyncError = null;
+    await sync.disconnect();
+    throw error;
   }
 }
 
 export async function clearData() {
-  await sync.disconnect();
+  await sync?.disconnect();
   await db.disconnectAndClear();
   await schemaManager.clear();
   await schemaManager.refreshSchema(db.database);
   if (connector.hasCredentials()) {
     const params = getParams();
-    await sync.connect({ params });
+    await sync?.connect({ params });
   }
 }
 
 export async function disconnect() {
-  await sync.disconnect();
+  await sync?.disconnect();
 }
 
 export async function signOut() {
   connector.clearCredentials();
+  await disconnect();
   await db.disconnectAndClear();
   await schemaManager.clear();
 }