+++ tools/powersynctests/e2e/jest.config.js
@@ -1,0 +1,21 @@
+/** @type {import('@jest/types').Config.InitialOptions} */
+module.exports = {
+  preset: 'ts-jest',
+  rootDir: '..',
+  testMatch: ['<rootDir>/e2e/**/*.test.ts'],
+  testTimeout: 120000,
+  maxWorkers: 1,
+  globalSetup: 'detox/runners/jest/globalSetup',
+  globalTeardown: 'detox/runners/jest/globalTeardown',
+  reporters: ['detox/runners/jest/reporter'],
+  testEnvironment: 'detox/runners/jest/testEnvironment',
+  verbose: true,
+  transform: {
+    '.+.[jt]sx?$': [
+      'babel-jest',
+      {
+        configFile: './babel.config.js',
+      },
+    ],
+  },
+};